added comments

+++ javaoop/exercises/_1/Course.java
@@ -1,9 +1,13 @@
 package javabasics.javaoop.exercises._1;
 
+// Created a new class called Course inside exercise._1 package
 public class Course {
+    // added variables courseName, maxStudents, qualityRatingOutOf10 with proper type
+
     String courseName = "";
-    int maxStudents = 0;
+    short maxStudents = 0;
     byte qualityRatingOutOf10 = 0;
+    // then added 2 other variables of my choice as the exercise request.
     String subject = "";
     String teacher = "";
